requirements.txt
Django==3.1.6
django-filter==2.4.0
django-oauth-toolkit==1.4.0
djangorestframework==3.12.2
sqlparse==0.4.1
gunicorn==20.0.4
